+++ packages/terra-site/src/examples/form-checkbox/HiddenLabelCheckboxExample.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import Checkbox from 'terra-form-checkbox/lib/Checkbox';
+
+const hiddenLabelCheckboxExample = () => (
+  <div>
+    <Checkbox labelText="Can you see me?" isLabelHidden />
+  </div>
+);
+
+export default hiddenLabelCheckboxExample;