+++ packages/terra-site/src/examples/form-checkbox/MultipleCheckboxesExample.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import Checkbox from 'terra-form-checkbox/lib/Checkbox';
+
+const multipleCheckboxesExample = () => (
+  <div>
+    <Checkbox id="first" labelText="First Checkbox" defaultChecked />
+    <Checkbox id="second" labelText="Second Checkbox" />
+    <Checkbox id="third" labelText="Third Checkbox" />
+  </div>
+);
+
+export default multipleCheckboxesExample;