+++ packages/terra-site/CHANGELOG.md
@@ -25,6 +25,7 @@ Unreleased
 1.18.0 - (November 16, 2017)
 ------------------
 ### Added
+* Add terra-form-checkbox components
 * Add terra-form-textarea components
 * Add terra-form-field components
 * Add terra-mock-theme styles